Implement death screen
Implement a death screen that occurs when player runs out of health or
big enemies upgrade to their final form.
Added another form for big enemies so that if is more obvious that they
are a big threat and can end the game. The upgraded form is bigger than
the original.